requirements.txt
airflow-dbt==0.4.0
apache-airflow-providers-airbyte
apache-airflow-providers-http
apache-airflow-providers-airbyte
apache-airflow-providers-slack
apache-airflow-providers-slack[http]